Fixed missing updates when props such as acceptInput change in isolation. This was because shouldComponentUpdate only compares emulatorState and inputStr. In fact there's no need to override shouldComponentUpdate.
Refactored ReactTerminal so that it inherits some props from ReactTerminalStateless.

Expose a new autoFocus prop. The autoFocus behaviour was introduced in 1.0.3 but was always on and there was no prop to control it. Now we've got a prop which defaults to true.

Fixed autoFocus so that it also applies when the component updates.

Updated docs and stories.