Use Styled Components as peer dependencies and do not bundle them
Currently Styled Components (along with React) were bundled in the final module bundle. This change addresses
this by moving `styled-components` as a peer dependency and specifying it as an external module to Webpack.
React and ReactDOM are also moved as external modules so that they are not present in the final bundle.

Additional changes
* Update Styled Components to the latest version
* Remove `styled-theming` dependency as it was not used
* Remove the `theme` default prop from components because it is being passed by the theme provider.
* Add basic Prettier config with a one rule to use single quote for strings (same as in ESLint config).